requirements.txt
nd2>=0.10.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tifffile>=2023.7.0 